move outsourced extensions into /exts
