+++ .flake8
@@ -87,6 +87,4 @@ per-file-ignores =
     extensions/*/__init__.py:ANN001,ANN201
     extensions/*/db.py:A003
     extensions/*/ext.py:ANN201
-
-    # from ext.py outsourced ext-files:
-    extensions/administration/sudo.py:ANN201
+    extensions/*/exts/*.py:ANN201
