+++ extensions/administration/exts/__init__.py
@@ -1,0 +1,1 @@
+from .sudo import *
